elliptic math makes my head hurt

+++ RawSocketTest/Payloads/Transform.cs
@@ -4,6 +4,10 @@ public class Transform
 {
     public ushort Length { get; set; }
     public TransformType Type { get; set; }
-    public ushort Id { get; set; }
+    
+    /// <summary>
+    /// This can be <see cref="EncryptionTypeId"/> or <see cref="DhId"/> depending on context
+    /// </summary>
+    public uint Id { get; set; }
     public List<Attribute> Attributes { get; set; } = new();
 }
